Complete WTA CLI boundary extraction
Move WT one-shot commands and delegate orchestration into focused CLI modules, leaving main.rs responsible for process setup and service-mode dispatch. Update the architecture refactor plan to mark the CLI boundary complete.


Copilot-Session: f0b2a188-f62a-48b7-b316-6fc31ed3880c
