+++ tools/wta/src/cli/mod.rs
@@ -1,4 +1,70 @@
 pub(crate) mod args;
+pub(crate) mod delegate;
 pub(crate) mod hooks;
 pub(crate) mod probes;
 pub(crate) mod sessions;
+pub(crate) mod wt;
+
+use anyhow::Result;
+
+use args::{Command, HooksAction, SessionsAction};
+
+pub(crate) async fn run(command: Command, json_mode: bool) -> Result<()> {
+    match command {
+        command @ (Command::Info
+        | Command::TestPipe
+        | Command::ListWindows
+        | Command::ListTabs { .. }
+        | Command::ListPanes { .. }
+        | Command::NewTab { .. }
+        | Command::SplitPane { .. }
+        | Command::CapturePane { .. }
+        | Command::KillPane { .. }
+        | Command::ActivePane
+        | Command::PaneStatus { .. }
+        | Command::WaitFor { .. }
+        | Command::PipeId
+        | Command::SetEnv { .. }
+        | Command::Listen { .. }) => wt::run(command, json_mode).await,
+        Command::ResolveCommand { token, shell } => {
+            let result = crate::resolve_command::resolve(&token, &shell).await;
+            if json_mode {
+                println!("{}", serde_json::to_string_pretty(&result)?);
+            } else {
+                println!("{}", crate::resolve_command::format_human(&result));
+            }
+            Ok(())
+        }
+        Command::Delegate {
+            prompt,
+            agent,
+            delegate_agent,
+            delegate_model,
+            cwd,
+        } => {
+            delegate::run(
+                prompt.as_deref(),
+                &agent,
+                delegate_agent.as_deref(),
+                delegate_model.as_deref(),
+                cwd.as_deref(),
+            )
+            .await
+        }
+        Command::Sessions { action } => match action {
+            SessionsAction::List { master, origin } => {
+                sessions::run_list(master, origin.to_filter(), json_mode).await
+            }
+        },
+        Command::Hooks { action } => match action {
+            HooksAction::Install { cli } => hooks::run_install(cli),
+            HooksAction::Status => hooks::run_status(json_mode),
+            HooksAction::Uninstall { cli } => hooks::run_uninstall(cli, json_mode),
+        },
+        Command::ProbeModels { agent } => probes::run_models(&agent).await,
+        Command::ProbeAgentSources { wsl_distro } => probes::run_agent_sources(&wsl_distro).await,
+        Command::ProbeSessions { agent } => probes::run_sessions(&agent).await,
+        Command::ProbeHostSessions { agent } => probes::run_host_sessions(&agent).await,
+        Command::ProbeWslSessions { cli } => probes::run_wsl_sessions(cli.as_deref()).await,
+    }
+}
